backend/requirements.txt
ruamel.yaml>=0.17.21
# ... existing requirements ... 